code refactoring to address pr comments
