+++ transformer_engine/common/gemm/ck_grouped_gemm/ck_grouped_gemm_common.h
@@ -6,8 +6,7 @@
 
 #pragma once
 
-#include <cuda.h>
-#include <cuda_runtime.h>
+#include <hip/hip_runtime.h>
 
 #include <array>
 #include <type_traits>
@@ -18,6 +17,7 @@
 #include "../../common.h"
 
 #include "ck_tile/core.hpp"
+#include "ck_tile/ops/epilogue.hpp"
 #include "ck_tile/ops/gemm.hpp"
 
 namespace transformer_engine {
@@ -25,6 +25,18 @@ namespace grouped_gemm {
 
 using RowMajor = ck_tile::tensor_layout::gemm::RowMajor;
 using ColMajor = ck_tile::tensor_layout::gemm::ColumnMajor;
+
+template <typename TileCfg>
+using GroupedGemmShape = ck_tile::TileGemmShape<
+    ck_tile::sequence<TileCfg::M_Tile, TileCfg::N_Tile, TileCfg::K_Tile>,
+    ck_tile::sequence<TileCfg::M_Warp, TileCfg::N_Warp, TileCfg::K_Warp>,
+    ck_tile::sequence<TileCfg::M_Warp_Tile, TileCfg::N_Warp_Tile, TileCfg::K_Warp_Tile>>;
+
+template <typename TileCfg>
+using GroupedGemmPartitioner = ck_tile::GemmSpatiallyLocalTilePartitioner<
+    GroupedGemmShape<TileCfg>,
+    TileCfg::TilePartitionerGroupNum,
+    TileCfg::TilePartitionerM01>;
 
 template <typename TEScalar> struct TETypeToCKType;
 template <> struct TETypeToCKType<transformer_engine::fp32>    { using type = float; };
@@ -42,6 +54,7 @@ struct EpilogueTraits {
   using DsLayout   = ck_tile::tuple<>;
   using ElemOp     = ck_tile::element_wise::PassThrough;
 };
+
 template <typename CType, typename CLayout>
 struct EpilogueTraits<CType, CLayout, true> {
   using DsDataType = ck_tile::tuple<CType>;
@@ -69,7 +82,7 @@ struct GroupedGemmRunContext {
 
     void* workspace = nullptr;
     size_t workspace_bytes = 0;
-    cudaStream_t stream = nullptr;
+    hipStream_t stream = nullptr;
 
     bool use_b_columnwise_data = false;
     bool accumulate = false;
@@ -87,22 +100,69 @@ static inline bool get_flat_2d_dims(const transformer_engine::Tensor& t,
   return true;
 }
 
-bool ck_tile_grouped_gemm_fp16_dispatch(DType a_dtype,
-                                       DType b_dtype,
-                                       DType d_dtype,
-                                       const GroupedGemmRunContext& ctx);
+// Extract GEMM dims from columnwise storage.
+// This path expects columnwise_data to already be normalized to a 2D layout.
+static inline bool get_columnwise_storage_2d_dims(
+    const transformer_engine::SimpleTensor& t,
+    int64_t& d0,
+    int64_t& d1) {
 
-bool ck_tile_grouped_gemm_fp8_dispatch(DType a_dtype,
-                                       DType b_dtype,
-                                       DType d_dtype,
-                                       const GroupedGemmRunContext& ctx);
+  if (t.shape.size() != 2) {
+    return false;
+  }
+
+  d0 = static_cast<int64_t>(t.shape[0]);
+  d1 = static_cast<int64_t>(t.shape[1]);
+  return true;
+}
+
+template <typename Kernel>
+static inline bool has_sufficient_workspace(const GroupedGemmRunContext& ctx) {
+  const size_t needed = Kernel::GetWorkSpaceSize(ctx.group_num);
+  if (!ctx.workspace || ctx.workspace_bytes < needed) {
+    NVTE_WARN("ck_tile_grouped_gemm: insufficient workspace for CK path. Needed bytes=", needed,
+              ", available bytes=", ctx.workspace_bytes, ". Falling back.");
+    return false;
+  }
+  return true;
+}
+
+template <typename Kernel, typename DescContainer>
+static inline bool launch_grouped_gemm_kernel(const DescContainer& descs,
+                                              const GroupedGemmRunContext& ctx,
+                                              const ck_tile::stream_config& stream_cfg) {
+  constexpr int kBlockPerCu = 1;
+
+  const dim3 blocks = Kernel::BlockSize();
+  const dim3 grids  = Kernel::GridSize(descs);
+  auto kargs = Kernel::MakeKargs(descs);
+
+  if (!Kernel::IsSupportedArgument(kargs)) {
+    NVTE_WARN("ck_tile_grouped_gemm: CK_Tile kernel arguments not supported for this config. "
+              "Falling back.");
+    return false;
+  }
+
+  NVTE_CHECK_CUDA(hipMemcpyAsync(ctx.workspace,
+                                  kargs.data(),
+                                  kargs.size() * sizeof(typename decltype(kargs)::value_type),
+                                  hipMemcpyHostToDevice,
+                                  ctx.stream));
+
+  ck_tile::launch_kernel(
+      stream_cfg, ck_tile::make_kernel<kBlockPerCu>(
+                      Kernel{}, grids, blocks, 0,
+                      ck_tile::cast_pointer_to_constant_address_space(ctx.workspace),
+                      ctx.group_num));
+  return true;
+}
 
 class RunnerInterface {
 public:
     virtual ~RunnerInterface() = default;
     virtual bool run(const ck_tile::stream_config& stream_cfg,
-                const GroupedGemmRunContext& ctx) = 0;
+                     const GroupedGemmRunContext& ctx) = 0;
 };
-                                                 
+
 }  // namespace grouped_gemm
 }  // namespace transformer_engine
