+++ transformer_engine/common/gemm/ck_grouped_gemm/ck_grouped_gemm_fp8.cpp
@@ -5,8 +5,14 @@
  ************************************************************************/
 
 #include "ck_grouped_gemm_common.h"
-#include "ck_grouped_gemm_fp8_impl.h"
-#include "common/util/cuda_runtime.h"
+#include "ck_grouped_gemm_fp8.h"
+
+#include <hip/hip_runtime.h>
+
+#include "ck_tile/ops/gemm_quant/kernel/grouped_gemm_quant_kernel.hpp"
+#include "ck_tile/ops/gemm_quant/pipeline/gemm_group_quant_utils.hpp"
+#include "ck_tile/ops/gemm_quant/pipeline/gemm_quant_pipeline_problem.hpp"
+#include "ck_tile/ops/gemm_quant/pipeline/tile_gemm_quant_traits.hpp"
 
 namespace transformer_engine {
 namespace grouped_gemm {
@@ -17,13 +23,272 @@ enum class GPUArch {
   UNKNOWN
 };
 
-static inline GPUArch detect_gpu_arch() {
-  int arch = cuda::sm_arch(0);
+struct TileCfg_128x128x128_16x16x128_2x2x1 {
+  static constexpr ck_tile::index_t M_Tile = 128;
+  static constexpr ck_tile::index_t N_Tile = 128;
+  static constexpr ck_tile::index_t K_Tile = 128;
 
-  if (arch == 94) {
+  static constexpr ck_tile::index_t M_Warp = 2;
+  static constexpr ck_tile::index_t N_Warp = 2;
+  static constexpr ck_tile::index_t K_Warp = 1;
+
+  static constexpr ck_tile::index_t M_Warp_Tile = 16;
+  static constexpr ck_tile::index_t N_Warp_Tile = 16;
+  static constexpr ck_tile::index_t K_Warp_Tile = 128;
+
+  static constexpr bool kPadM = false;
+  static constexpr bool kPadN = false;
+  static constexpr bool kPadK = false;
+
+  static constexpr bool DoubleSmemBuffer = false;
+
+  static constexpr ck_tile::index_t TilePartitionerGroupNum = 16;
+  static constexpr ck_tile::index_t TilePartitionerM01 = 8;
+};
+
+// gfx950 device compilation cannot instantiate the literal 32x32x16 FP8 tile
+// configuration due to an unsupported warp GEMM dispatcher configuration.
+// See: ck_tile/ops/gemm/warp/warp_gemm_dispatcher.hpp for supported variants.
+//
+// To preserve the existing type name in shared template code, this struct
+// inherits from the gfx950-safe 16x16x128 configuration in the gfx950 device
+// compilation path, effectively reusing those parameters without redefining them.
+//
+// In all other compilation paths, the struct overrides the relevant fields to
+// provide the intended 32x32x16 configuration.
+#if defined(__gfx950__)
+struct TileCfg_128x128x128_32x32x16_2x2x1
+    : TileCfg_128x128x128_16x16x128_2x2x1 {};
+#else
+struct TileCfg_128x128x128_32x32x16_2x2x1
+    : TileCfg_128x128x128_16x16x128_2x2x1 {
+  static constexpr ck_tile::index_t M_Warp_Tile = 32;
+  static constexpr ck_tile::index_t N_Warp_Tile = 32;
+  static constexpr ck_tile::index_t K_Warp_Tile = 16;
+
+  static constexpr ck_tile::index_t TilePartitionerGroupNum = 8;
+  static constexpr ck_tile::index_t TilePartitionerM01 = 4;
+};
+#endif
+
+// FP8 currently supports overwrite only.
+// Preserve MemOp here for a future accumulate path.
+template <typename AType,
+          typename BType,
+          typename CType,
+          typename ALayout,
+          typename BLayout,
+          typename CLayout,
+          typename TileCfg,
+          ck_tile::memory_operation_enum MemOp,
+          typename AccType = float>
+class QuantGroupedGemmRunner : public RunnerInterface {
+ public:
+  static constexpr ck_tile::QuantType QuantMode = ck_tile::QuantType::TensorQuant;
+
+  using GemmShape = GroupedGemmShape<TileCfg>;
+  using Partitioner = GroupedGemmPartitioner<TileCfg>;
+
+  using AQLayout = RowMajor;
+  using BQLayout = RowMajor;
+
+  using UniversalTraits =
+      ck_tile::TileGemmQuantTraits<TileCfg::kPadM,
+                                   TileCfg::kPadN,
+                                   TileCfg::kPadK,
+                                   false,
+                                   false,
+                                   false,
+                                   ALayout,
+                                   BLayout,
+                                   CLayout,
+                                   QuantMode,
+                                   AQLayout,
+                                   BQLayout,
+                                   false,
+                                   TileCfg::DoubleSmemBuffer,
+                                   false>;
+
+  using Problem =
+      ck_tile::GemmRowColTensorQuantPipelineProblem<AType,
+                                                    BType,
+                                                    AccType,
+                                                    AccType,
+                                                    GemmShape,
+                                                    UniversalTraits,
+                                                    false,
+                                                    AccType>;
+
+  using Pipeline = ck_tile::GemmPipelineAgBgCrCompV3<Problem>;
+
+  using Epilogue =
+      ck_tile::CShuffleEpilogue<ck_tile::CShuffleEpilogueProblem<
+          AType,
+          BType,
+          ck_tile::tuple<>,
+          AccType,
+          CType,
+          ck_tile::tuple<>,
+          CLayout,
+          ck_tile::element_wise::PassThrough,
+          Partitioner::MPerBlock,
+          Partitioner::NPerBlock,
+          TileCfg::M_Warp,
+          TileCfg::N_Warp,
+          TileCfg::M_Warp_Tile,
+          TileCfg::N_Warp_Tile,
+          TileCfg::K_Warp_Tile,
+          Problem::TransposeC>>;
+
+  using Kernel =
+      ck_tile::QuantGroupedGemmKernel<Partitioner, Pipeline, Epilogue, QuantMode>;
+  using HostArgs = ck_tile::QuantGroupedGemmHostArgs;
+
+ public:
+  static std::vector<HostArgs> build_descs(const GroupedGemmRunContext& ctx) {
+    if (!has_sufficient_workspace<Kernel>(ctx)) {
+      return {};
+    }
+
+    std::vector<HostArgs> descs;
+    descs.reserve(ctx.group_num);
+
+    for (int i = 0; i < ctx.group_num; ++i) {
+      const transformer_engine::Tensor* const A_te =
+          transformer_engine::convertNVTETensorCheck(ctx.A[i]);
+      const transformer_engine::Tensor* const B_te =
+          transformer_engine::convertNVTETensorCheck(ctx.B[i]);
+      transformer_engine::Tensor* D_te =
+          transformer_engine::convertNVTETensorCheck(ctx.D[i]);
+
+      const auto& a = data_view(*A_te);
+      const auto& d = data_view(*D_te);
+
+      const transformer_engine::SimpleTensor* b_src = nullptr;
+      if (ctx.use_b_columnwise_data) {
+        if (!B_te->has_columnwise_data()) {
+          NVTE_ERROR("ck_tile_grouped_gemm: ctx.use_b_columnwise_data=true but "
+                     "columnwise_data is absent.");
+        }
+        b_src = &B_te->columnwise_data;
+      } else {
+        b_src = &B_te->data;
+      }
+
+      const auto& b = *b_src;
+
+      int64_t Ad0 = 0, Ad1 = 0, Dd0 = 0, Dd1 = 0;
+      if (!get_flat_2d_dims(*A_te, Ad0, Ad1) ||
+          !get_flat_2d_dims(*D_te, Dd0, Dd1)) {
+        NVTE_ERROR("ck_tile_grouped_gemm: expected A and D to be rank>=2.");
+      }
+
+      if (b.shape.size() < 2) {
+        NVTE_ERROR("ck_tile_grouped_gemm: expected chosen B source to be rank>=2.");
+      }
+
+      int64_t Bd0 = static_cast<int64_t>(b.shape[b.shape.size() - 2]);
+      int64_t Bd1 = static_cast<int64_t>(b.shape[b.shape.size() - 1]);
+
+      const int64_t M = ctx.transA ? Ad1 : Ad0;
+      const int64_t K = ctx.transA ? Ad0 : Ad1;
+      const int64_t N = ctx.transB ? Bd0 : Bd1;
+      const int64_t Kb = ctx.transB ? Bd1 : Bd0;
+
+      if (Kb != K) {
+        NVTE_ERROR("ck_tile_grouped_gemm: K mismatch between A and B in group ",
+                   i,
+                   ". op(A)=",
+                   M,
+                   "x",
+                   K,
+                   " op(B)=",
+                   Kb,
+                   "x",
+                   N,
+                   " raw A=",
+                   Ad0,
+                   "x",
+                   Ad1,
+                   " raw B=",
+                   Bd0,
+                   "x",
+                   Bd1,
+                   " use_b_columnwise_data=",
+                   static_cast<int>(ctx.use_b_columnwise_data),
+                   " transA=",
+                   static_cast<int>(ctx.transA),
+                   " transB=",
+                   static_cast<int>(ctx.transB));
+      }
+
+      if (Dd0 != M || Dd1 != N) {
+        NVTE_ERROR("ck_tile_grouped_gemm: D shape mismatch in group ",
+                   i,
+                   ". D=",
+                   Dd0,
+                   "x",
+                   Dd1,
+                   ", expected=",
+                   M,
+                   "x",
+                   N);
+      }
+
+      const ck_tile::index_t stride_A = static_cast<ck_tile::index_t>(Ad1);
+      const ck_tile::index_t stride_B = static_cast<ck_tile::index_t>(Bd1);
+      const ck_tile::index_t stride_E = static_cast<ck_tile::index_t>(Dd1);
+
+      ck_tile::index_t AQK = 1;
+      ck_tile::index_t BQK = 1;
+      ck_tile::index_t stride_AQ = 1;
+      ck_tile::index_t stride_BQ = 1;
+
+      const auto& aq = scale_inv_view(*A_te);
+      const auto& bq = scale_inv_view(*B_te);
+
+      descs.emplace_back(a.dptr,
+                         b.dptr,
+                         d.dptr,
+                         aq.dptr,
+                         bq.dptr,
+                         1,
+                         M,
+                         N,
+                         K,
+                         AQK,
+                         BQK,
+                         stride_A,
+                         stride_B,
+                         stride_E,
+                         stride_AQ,
+                         stride_BQ);
+    }
+
+    return descs;
+  }
+
+  bool run(const ck_tile::stream_config& stream_cfg,
+           const GroupedGemmRunContext& ctx) override {
+    auto descs = build_descs(ctx);
+    if (descs.empty()) {
+      return false;
+    }
+    return launch_grouped_gemm_kernel<Kernel>(descs, ctx, stream_cfg);
+  }
+};
+
+static inline GPUArch detect_gpu_arch() {
+  int device = 0;
+  HIP_CHECK_ERROR(hipGetDevice(&device));
+
+  hipDeviceProp_t props{};
+  HIP_CHECK_ERROR(hipGetDeviceProperties(&props, device));
+
+  if (props.major == 9 && props.minor == 4) {
     return GPUArch::GFX942;
   }
-  if (arch == 95) {
+  if (props.major == 9 && props.minor == 5) {
     return GPUArch::GFX950;
   }
   return GPUArch::UNKNOWN;
@@ -43,8 +308,9 @@ struct FP8TileCfg<GPUArch::GFX950> {
 };
 
 template <GPUArch Arch, typename teA, typename teB, typename ALayout, typename BLayout>
-static std::unique_ptr<RunnerInterface> make_fp8_runner_typed(DType d_dtype,
-                                                       const GroupedGemmRunContext& ctx) {
+static std::unique_ptr<RunnerInterface> make_fp8_runner_typed(
+    DType d_dtype,
+    const GroupedGemmRunContext& ctx) {
   std::unique_ptr<RunnerInterface> runner = nullptr;
 
   using AType = typename TETypeToCKType<teA>::type;
@@ -54,10 +320,14 @@ static std::unique_ptr<RunnerInterface> make_fp8_runner_typed(DType d_dtype,
 
   TRANSFORMER_ENGINE_TYPE_SWITCH_NON_FP8ONLY(d_dtype, d_te_type, {
     using CType = typename TETypeToCKType<d_te_type>::type;
-    using Runner = QuantGroupedGemmRunner<
-        AType, BType, CType,
-        ALayout, BLayout, CTypeLayout,
-        TileCfg, ck_tile::memory_operation_enum::set>;
+    using Runner = QuantGroupedGemmRunner<AType,
+                                          BType,
+                                          CType,
+                                          ALayout,
+                                          BLayout,
+                                          CTypeLayout,
+                                          TileCfg,
+                                          ck_tile::memory_operation_enum::set>;
     runner = std::make_unique<Runner>();
   });
 
@@ -65,48 +335,50 @@ static std::unique_ptr<RunnerInterface> make_fp8_runner_typed(DType d_dtype,
 }
 
 template <GPUArch Arch>
-static std::unique_ptr<RunnerInterface> make_fp8_runner_impl(DType a_dtype,
-                                                      DType b_dtype,
-                                                      DType d_dtype,
-                                                      const GroupedGemmRunContext& ctx) {
+static std::unique_ptr<RunnerInterface> make_fp8_runner_impl(
+    DType a_dtype,
+    DType b_dtype,
+    DType d_dtype,
+    const GroupedGemmRunContext& ctx) {
+  TRANSFORMER_ENGINE_SWITCH_CONDITION(ctx.transA, kTransA, {
+    using ALayout = std::conditional_t<kTransA, ColMajor, RowMajor>;
 
-    TRANSFORMER_ENGINE_SWITCH_CONDITION(ctx.transA, kTransA, {
-        using ALayout = std::conditional_t<kTransA, ColMajor, RowMajor>;
+    TRANSFORMER_ENGINE_SWITCH_CONDITION(ctx.transB, kTransB, {
+      using BLayout = std::conditional_t<kTransB, ColMajor, RowMajor>;
 
-        TRANSFORMER_ENGINE_SWITCH_CONDITION(ctx.transB, kTransB, {
-            using BLayout = std::conditional_t<kTransB, ColMajor, RowMajor>;
-
-            TRANSFORMER_ENGINE_TYPE_SWITCH_FP8ONLY(a_dtype, a_type, {
-                TRANSFORMER_ENGINE_TYPE_SWITCH_FP8ONLY(b_dtype, b_type, {
-                    return make_fp8_runner_typed<Arch, a_type, b_type, ALayout, BLayout>(
-                        d_dtype, ctx);
-                });
-            });
+      TRANSFORMER_ENGINE_TYPE_SWITCH_FP8ONLY(a_dtype, a_type, {
+        TRANSFORMER_ENGINE_TYPE_SWITCH_FP8ONLY(b_dtype, b_type, {
+          return make_fp8_runner_typed<Arch, a_type, b_type, ALayout, BLayout>(
+              d_dtype, ctx);
         });
+      });
     });
+  });
 
-    return nullptr;
+  return nullptr;
 }
 
-static inline std::unique_ptr<RunnerInterface> make_fp8_runner_gfx942(DType a_dtype,
-                                                        DType b_dtype,
-                                                        DType d_dtype,
-                                                        const GroupedGemmRunContext& ctx) {
+static inline std::unique_ptr<RunnerInterface> make_fp8_runner_gfx942(
+    DType a_dtype,
+    DType b_dtype,
+    DType d_dtype,
+    const GroupedGemmRunContext& ctx) {
   return make_fp8_runner_impl<GPUArch::GFX942>(a_dtype, b_dtype, d_dtype, ctx);
 }
 
-static inline std::unique_ptr<RunnerInterface> make_fp8_runner_gfx950(DType a_dtype,
-                                                        DType b_dtype,
-                                                        DType d_dtype,
-                                                        const GroupedGemmRunContext& ctx) {
+static inline std::unique_ptr<RunnerInterface> make_fp8_runner_gfx950(
+    DType a_dtype,
+    DType b_dtype,
+    DType d_dtype,
+    const GroupedGemmRunContext& ctx) {
   return make_fp8_runner_impl<GPUArch::GFX950>(a_dtype, b_dtype, d_dtype, ctx);
 }
 
-static std::unique_ptr<RunnerInterface> 
-make_fp8_runner(DType a_dtype,
-                DType b_dtype,
-                DType d_dtype,
-                const GroupedGemmRunContext& ctx) {
+static std::unique_ptr<RunnerInterface> make_fp8_runner(
+    DType a_dtype,
+    DType b_dtype,
+    DType d_dtype,
+    const GroupedGemmRunContext& ctx) {
   switch (detect_gpu_arch()) {
     case GPUArch::GFX942:
       return make_fp8_runner_gfx942(a_dtype, b_dtype, d_dtype, ctx);
@@ -123,10 +395,12 @@ bool ck_tile_grouped_gemm_fp8_dispatch(DType a_dtype,
                                        DType d_dtype,
                                        const GroupedGemmRunContext& ctx) {
   const ck_tile::stream_config s{ctx.stream};
+
   auto runner = make_fp8_runner(a_dtype, b_dtype, d_dtype, ctx);
   if (!runner) {
     return false;
   }
+
   return runner->run(s, ctx);
 }
