--- transformer_engine/common/gemm/ck_grouped_gemm/ck_grouped_gemm_fp8_impl.h
@@ -1,260 +1,0 @@
-/*************************************************************************
- * Copyright (c) 2026, Advanced Micro Devices, Inc. All rights reserved.
- *
- * License for AMD contributions = MIT. See LICENSE for more information
- ************************************************************************/
-
-#pragma once
-#include "ck_grouped_gemm_common.h"
-
-#include <array>
-#include <type_traits>
-#include <vector>
-#include <memory>
-
-#include <transformer_engine/transformer_engine.h>
-
-#include "ck_tile/core.hpp"
-#include "ck_tile/ops/epilogue.hpp"
-#include "ck_tile/ops/gemm.hpp"
-
-#include "ck_tile/ops/gemm_quant/pipeline/gemm_group_quant_utils.hpp"
-#include "ck_tile/ops/gemm_quant/pipeline/tile_gemm_quant_traits.hpp"
-#include "ck_tile/ops/gemm_quant/pipeline/gemm_quant_pipeline_problem.hpp"
-#include "ck_tile/ops/gemm_quant/kernel/grouped_gemm_quant_kernel.hpp"
-
-namespace transformer_engine {
-namespace grouped_gemm {
-
-struct TileCfg_128x128x128_16x16x128_2x2x1 {
-  static constexpr ck_tile::index_t M_Tile = 128;
-  static constexpr ck_tile::index_t N_Tile = 128;
-  static constexpr ck_tile::index_t K_Tile = 128;
-
-  static constexpr ck_tile::index_t M_Warp = 2;
-  static constexpr ck_tile::index_t N_Warp = 2;
-  static constexpr ck_tile::index_t K_Warp = 1;
-
-  static constexpr ck_tile::index_t M_Warp_Tile = 16;
-  static constexpr ck_tile::index_t N_Warp_Tile = 16;
-  static constexpr ck_tile::index_t K_Warp_Tile = 128;
-
-  static constexpr bool kPadM = false;
-  static constexpr bool kPadN = false;
-  static constexpr bool kPadK = false;
-
-  static constexpr bool DoubleSmemBuffer = false;
-
-  static constexpr ck_tile::index_t TilePartitionerGroupNum = 16;
-  static constexpr ck_tile::index_t TilePartitionerM01      = 8;
-};
-
-// gfx950 device compilation cannot instantiate the literal 32x32x16 FP8 tile
-// configuration due to an unsupported warp GEMM dispatcher configuration.
-// See: ck_tile/ops/gemm/warp/warp_gemm_dispatcher.hpp for supported variants.
-//
-// To preserve the existing type name in shared template code, this struct
-// inherits from the gfx950-safe 16x16x128 configuration in the gfx950 device
-// compilation path, effectively reusing those parameters without redefining them.
-//
-// In all other compilation paths, the struct overrides the relevant fields to
-// provide the intended 32x32x16 configuration.
-#if defined(__gfx950__)
-struct TileCfg_128x128x128_32x32x16_2x2x1
-    : TileCfg_128x128x128_16x16x128_2x2x1 {
-};
-#else
-struct TileCfg_128x128x128_32x32x16_2x2x1
-    : TileCfg_128x128x128_16x16x128_2x2x1 {
-  static constexpr ck_tile::index_t M_Warp_Tile = 32;
-  static constexpr ck_tile::index_t N_Warp_Tile = 32;
-  static constexpr ck_tile::index_t K_Warp_Tile = 16;
-
-  static constexpr ck_tile::index_t TilePartitionerGroupNum = 8;
-  static constexpr ck_tile::index_t TilePartitionerM01      = 4;
-};
-#endif
-
-template <typename AType, typename BType, typename CType,
-          typename ALayout, typename BLayout, typename CLayout,
-          typename TileCfg,
-          ck_tile::memory_operation_enum MemOp,
-          typename AccType = float>
-class QuantGroupedGemmRunner : public RunnerInterface {
-public:
-    static constexpr ck_tile::QuantType QuantMode = ck_tile::QuantType::TensorQuant;
-
-    using GemmShape = ck_tile::TileGemmShape<
-        ck_tile::sequence<TileCfg::M_Tile, TileCfg::N_Tile, TileCfg::K_Tile>,
-        ck_tile::sequence<TileCfg::M_Warp, TileCfg::N_Warp, TileCfg::K_Warp>,
-        ck_tile::sequence<TileCfg::M_Warp_Tile, TileCfg::N_Warp_Tile, TileCfg::K_Warp_Tile>>;
-
-    using Partitioner = ck_tile::GemmSpatiallyLocalTilePartitioner<
-        GemmShape, TileCfg::TilePartitionerGroupNum, TileCfg::TilePartitionerM01>;
-
-    using AQLayout = RowMajor;
-    using BQLayout = RowMajor;
-
-    using UniversalTraits =
-        ck_tile::TileGemmQuantTraits<
-            TileCfg::kPadM, TileCfg::kPadN, TileCfg::kPadK,
-            false, false, false, ALayout, BLayout, CLayout,
-            QuantMode, AQLayout, BQLayout,
-            false, TileCfg::DoubleSmemBuffer, false>;
-
-    using Problem = ck_tile::GemmRowColTensorQuantPipelineProblem<
-            AType, BType, AccType,
-            AccType, GemmShape, UniversalTraits,
-            false, AccType>;
-
-    using Pipeline = ck_tile::GemmPipelineAgBgCrCompV3<Problem>;
-
-    using Epilogue = ck_tile::CShuffleEpilogue<
-        ck_tile::CShuffleEpilogueProblem<
-            AType, BType, ck_tile::tuple<>, AccType,
-            CType, ck_tile::tuple<>, CLayout,
-            ck_tile::element_wise::PassThrough,
-            Partitioner::MPerBlock, Partitioner::NPerBlock,
-            TileCfg::M_Warp, TileCfg::N_Warp,
-            TileCfg::M_Warp_Tile, TileCfg::N_Warp_Tile, TileCfg::K_Warp_Tile,
-            Problem::TransposeC>>;
-
-    using Kernel = ck_tile::QuantGroupedGemmKernel<Partitioner, Pipeline, Epilogue, QuantMode>;
-    using HostArgs = ck_tile::QuantGroupedGemmHostArgs;
-
-public:
-    static std::vector<HostArgs> build_descs(const GroupedGemmRunContext& ctx) {
-       const size_t needed = Kernel::GetWorkSpaceSize(ctx.group_num);
-        if (!ctx.workspace || ctx.workspace_bytes < needed) {
-            NVTE_WARN("ck_tile_grouped_gemm: insufficient workspace for CK path. Needed bytes=", needed,
-                    ", available bytes=", ctx.workspace_bytes, ". Falling back.");
-            return {};
-        }
-
-      std::vector<HostArgs> descs;
-      descs.reserve(ctx.group_num);
-
-      for (int i = 0; i < ctx.group_num; ++i) {
-          const transformer_engine::Tensor* const A_te =
-              transformer_engine::convertNVTETensorCheck(ctx.A[i]);
-          const transformer_engine::Tensor* const B_te =
-              transformer_engine::convertNVTETensorCheck(ctx.B[i]);
-          transformer_engine::Tensor* D_te =
-              transformer_engine::convertNVTETensorCheck(ctx.D[i]);
-
-          const auto& a = data_view(*A_te);
-          const auto& d = data_view(*D_te);
-
-          const transformer_engine::SimpleTensor* b_src = nullptr;
-          if (ctx.use_b_columnwise_data) {
-              if (!B_te->has_columnwise_data()) {
-                  NVTE_ERROR("ck_tile_grouped_gemm: ctx.use_b_columnwise_data=true but columnwise_data is absent.");
-              }
-              b_src = &B_te->columnwise_data;
-          } else {
-              b_src = &B_te->data;
-          }
-
-          const auto& b = *b_src;
-
-          int64_t Ad0 = 0, Ad1 = 0, Dd0 = 0, Dd1 = 0;
-          if (!get_flat_2d_dims(*A_te, Ad0, Ad1) ||
-              !get_flat_2d_dims(*D_te, Dd0, Dd1)) {
-              NVTE_ERROR("ck_tile_grouped_gemm: expected A and D to be rank>=2.");
-          }
-
-          if (b.shape.size() < 2) {
-              NVTE_ERROR("ck_tile_grouped_gemm: expected chosen B source to be rank>=2.");
-          }
-
-          int64_t Bd0 = static_cast<int64_t>(b.shape[b.shape.size() - 2]);
-          int64_t Bd1 = static_cast<int64_t>(b.shape[b.shape.size() - 1]);
-
-          const int64_t M  = ctx.transA ? Ad1 : Ad0;
-          const int64_t K  = ctx.transA ? Ad0 : Ad1;
-          const int64_t N  = ctx.transB ? Bd0 : Bd1;
-          const int64_t Kb = ctx.transB ? Bd1 : Bd0;
-
-          if (Kb != K) {
-              NVTE_ERROR("ck_tile_grouped_gemm: K mismatch between A and B in group ", i,
-                         ". op(A)=", M, "x", K,
-                         " op(B)=", Kb, "x", N,
-                         " raw A=", Ad0, "x", Ad1,
-                         " raw B=", Bd0, "x", Bd1,
-                         " use_b_columnwise_data=", static_cast<int>(ctx.use_b_columnwise_data),
-                         " transA=", static_cast<int>(ctx.transA),
-                         " transB=", static_cast<int>(ctx.transB));
-          }
-
-          if (Dd0 != M || Dd1 != N) {
-              NVTE_ERROR("ck_tile_grouped_gemm: D shape mismatch in group ", i,
-                         ". D=", Dd0, "x", Dd1,
-                         ", expected=", M, "x", N);
-          }
-
-          const ck_tile::index_t stride_A = static_cast<ck_tile::index_t>(Ad1);
-          const ck_tile::index_t stride_B = static_cast<ck_tile::index_t>(Bd1);
-          const ck_tile::index_t stride_E = static_cast<ck_tile::index_t>(Dd1);
-
-          ck_tile::index_t AQK       = 1;
-          ck_tile::index_t BQK       = 1;
-          ck_tile::index_t stride_AQ = 1;
-          ck_tile::index_t stride_BQ = 1;
-
-          const auto& aq = scale_inv_view(*A_te);
-          const auto& bq = scale_inv_view(*B_te);
-
-          descs.emplace_back(
-              a.dptr,
-              b.dptr,
-              d.dptr,
-              aq.dptr,
-              bq.dptr,
-              1,
-              M,
-              N,
-              K,
-              AQK,
-              BQK,
-              stride_A,
-              stride_B,
-              stride_E,
-              stride_AQ,
-              stride_BQ);
-        }
-
-        return descs;
-    }
-
-    bool run(const ck_tile::stream_config& stream_cfg,
-             const GroupedGemmRunContext& ctx) override {
-        auto descs = build_descs(ctx);
-
-        constexpr int kBlockPerCu = 1;
-        const dim3 blocks = Kernel::BlockSize();
-        const dim3 grids  = Kernel::GridSize(descs);
-        auto kargs = Kernel::MakeKargs(descs);
-
-        if (!Kernel::IsSupportedArgument(kargs)) {
-            NVTE_WARN("ck_tile_grouped_gemm: CK_Tile kernel arguments not supported for this config. "
-                    "Falling back.");
-            return false;
-        }
-
-        NVTE_CHECK_CUDA(cudaMemcpyAsync(ctx.workspace,
-                                       kargs.data(),
-                                       kargs.size() * sizeof(typename decltype(kargs)::value_type),
-                                       cudaMemcpyHostToDevice,
-                                       ctx.stream));
-
-        ck_tile::launch_kernel(
-            stream_cfg, ck_tile::make_kernel<kBlockPerCu>(
-                            Kernel{}, grids, blocks, 0,
-                            ck_tile::cast_pointer_to_constant_address_space(ctx.workspace),
-                            ctx.group_num));
-        return true;
-    }
-};
-
-}  // namespace grouped_gemm
-}  // namespace transformer_engine
