+++ transformer_engine/common/include/transformer_engine/transformer_engine.h
@@ -423,6 +423,7 @@ enum class DType {
   kNumTypes
 };
 
+#ifdef USE_ROCM
 /*! \brief Check if TE datatype is FP16
  *
  * Return true if TE datatype is FP16
@@ -431,7 +432,7 @@ enum class DType {
 inline bool is_fp16_dtype(const DType t) {
   return t == DType::kFloat16 || t == DType::kBFloat16;
 }
-
+#endif
 
 /*! \brief Check if TE datatype is FP8
  *
